chore(core): context aware welcome page
